Remove ghost images on drag in carousels
Add the draggable attribute to all img and anchor tags in carousels and set it to false.
This prevents the urls in anchor tags and images in img tags from being dragged out into
ghost images and preventing the swiping functionality on desktop when user clicks on the image
with a mouse to drag.